Created test for move with decimals
by [Dara Keon (desktop)] in [2014-12-30 00:36:33 (-02:00)] - hg hash {39f838f69c16}

+++ site/DFM.Tests/BusinessLogic/BaseStep.cs
@@ -347,7 +347,7 @@ namespace DFM.Tests.BusinessLogic
             var newDetail = new Detail { Description = detailData["Description"] };
 
             if (!String.IsNullOrEmpty(detailData["Value"]))
-                newDetail.Value = Int32.Parse(detailData["Value"]);
+                newDetail.Value = Double.Parse(detailData["Value"]);
 
             if (!String.IsNullOrEmpty(detailData["Amount"]))
                 newDetail.Amount = Int16.Parse(detailData["Amount"]);
